Upgrade game module to use 64-bit timestamps

+++ source/game/ai/ai_base_ai.h
@@ -74,7 +74,7 @@ protected:
 
 	inline bool IsCloseToReachStart() { return distanceToNextReachStart < 24.0f; };
 
-	unsigned blockedTimeout;
+	int64_t blockedTimeout;
 
 	float aiYawSpeed, aiPitchSpeed;
 	float oldYawAbsDiff, oldPitchAbsDiff;
