+++ source/game/g_gametypes.cpp
@@ -509,7 +509,7 @@ bool G_Match_CheckExtendPlayTime( void ) {
 
 				G_PrintMsg( NULL, "Match tied. Timelimit extended by %i minutes!\n", g_match_extendedtime->integer );
 				G_CenterPrintFormatMsg( NULL, "%s MINUTE OVERTIME\n", va( "%i", g_match_extendedtime->integer ), NULL );
-				gs.gameState.stats[GAMESTAT_MATCHDURATION] = (unsigned int)( ( fabs( g_match_extendedtime->value ) * 60 ) * 1000 );
+				gs.gameState.stats[GAMESTAT_MATCHDURATION] = (int64_t)( ( fabs( g_match_extendedtime->value ) * 60 ) * 1000 );
 			} else {
 				G_AnnouncerSound( NULL, trap_SoundIndex( va( S_ANNOUNCER_OVERTIME_SUDDENDEATH_1_to_2, ( rand() & 1 ) + 1 ) ), GS_MAX_TEAMS, true, NULL );
 				G_PrintMsg( NULL, "Match tied. Sudden death!\n" );
@@ -749,7 +749,7 @@ void G_Match_LaunchState( int matchState ) {
 			level.forceStart = false;
 
 			gs.gameState.stats[GAMESTAT_MATCHSTATE] = MATCH_STATE_WARMUP;
-			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (unsigned int)( fabs( g_warmup_timelimit->value * 60 ) * 1000 );
+			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (int64_t)( fabs( g_warmup_timelimit->value * 60 ) * 1000 );
 			gs.gameState.stats[GAMESTAT_MATCHSTART] = game.serverTime;
 
 			// race has playtime in warmup too, so flag the matchmaker about this
@@ -765,7 +765,7 @@ void G_Match_LaunchState( int matchState ) {
 			advance_queue = true;
 
 			gs.gameState.stats[GAMESTAT_MATCHSTATE] = MATCH_STATE_COUNTDOWN;
-			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (unsigned int)( fabs( g_countdown_time->value ) * 1000 );
+			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (int64_t)( fabs( g_countdown_time->value ) * 1000 );
 			gs.gameState.stats[GAMESTAT_MATCHSTART] = game.serverTime;
 
 			break;
@@ -779,7 +779,7 @@ void G_Match_LaunchState( int matchState ) {
 			level.forceStart = false;
 
 			gs.gameState.stats[GAMESTAT_MATCHSTATE] = MATCH_STATE_PLAYTIME;
-			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (unsigned int)( fabs( 60 * g_timelimit->value ) * 1000 );
+			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (int64_t)( fabs( 60 * g_timelimit->value ) * 1000 );
 			gs.gameState.stats[GAMESTAT_MATCHSTART] = game.serverTime;
 
 			// request a new match UUID
@@ -795,7 +795,7 @@ void G_Match_LaunchState( int matchState ) {
 		case MATCH_STATE_POSTMATCH:
 		{
 			gs.gameState.stats[GAMESTAT_MATCHSTATE] = MATCH_STATE_POSTMATCH;
-			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (unsigned int)fabs( g_postmatch_timelimit->value * 1000 ); // postmatch time in seconds
+			gs.gameState.stats[GAMESTAT_MATCHDURATION] = (int64_t)fabs( g_postmatch_timelimit->value * 1000 ); // postmatch time in seconds
 			gs.gameState.stats[GAMESTAT_MATCHSTART] = game.serverTime;
 
 			G_Timeout_Reset();
